Skapat fil docker-compose.yml och även klass "Ownership". Även fixat ett typo i mina repon, hade angivit Integer istället för Long.

+++ Padel/src/main/java/com/strom/wigellPadel/repositories/BookingRepo.java
@@ -3,5 +3,5 @@ package com.strom.wigellPadel.repositories;
 import com.strom.wigellPadel.entities.Booking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookingRepo extends JpaRepository<Booking, Integer> {
+public interface BookingRepo extends JpaRepository<Booking, Long> {
 }
