+++ Padel/src/main/java/com/strom/wigellPadel/repositories/CustomerRepo.java
@@ -3,5 +3,5 @@ package com.strom.wigellPadel.repositories;
 import com.strom.wigellPadel.entities.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepo extends JpaRepository<Customer, Integer> {
+public interface CustomerRepo extends JpaRepository<Customer, Long> {
 }
